finetune

+++ src/finetune.py
@@ -51,8 +51,14 @@ def finetune(subtask:str, shortcut_model_name:str):
         bnb_4bit_use_double_quant=True,
         bnb_4bit_compute_dtype=torch.float16,
     )
-    if shortcut_model_name == "phi_3_mini": model = AutoModelForCausalLM.from_pretrained(full_model_name, trust_remote_code=True, quantization_config=bnb_config, device_map="auto", attn_implementation="flash_attention_2")
-    else: model = AutoModelForCausalLM.from_pretrained(full_model_name, trust_remote_code=True, quantization_config=bnb_config, device_map="auto")
+    # model = AutoModelForCausalLM.from_pretrained(
+    #     full_model_name,
+    #     quantization_config=bnb_config,
+    #     device_map="auto",
+    #     trust_remote_code=True
+    # )
+    if shortcut_model_name == "phi_3_mini": model = AutoModelForCausalLM.from_pretrained(full_model_name, trust_remote_code=True, torch_dtype=torch.float16, quantization_config=bnb_config, device_map="auto", attn_implementation="flash_attention_2")
+    else: model = AutoModelForCausalLM.from_pretrained(full_model_name, trust_remote_code=True, torch_dtype=torch.float16, quantization_config=bnb_config, device_map="auto")
     model.config.use_cache = False
     model = prepare_model_for_kbit_training(model)
